Prepare to move the script-related code back to atlastool.cpp
